refactor(api-file-manager-server): DI-native server upload GraphQL schema factory
Same migration as api-file-manager-s3 #5483: legacy createGraphQLSchemaPlugin
(service-locating context.container.resolve in resolvers, addLegacyResolvers) ->
ServerGraphQLSchema CoreGraphQLSchemaFactory using builder.addResolver with declared
dependencies (IdentityContext / GetSettingsUseCase / GetUploadPayloadUseCase /
Create+CompleteMultiPartUploadUseCase). Behavior-preserving.
